Extract type registration to own module and include in Node, Document, and Text.
