+++ lib/tip_tap/document.rb
@@ -1,19 +1,12 @@
 # frozen_string_literal: true
 
-require "tip_tap/html_renderable"
-require "tip_tap/json_renderable"
-require "tip_tap/plain_text_renderable"
-require "tip_tap/has_content"
+require "tip_tap/registerable"
+require "tip_tap/node"
 
 # This is the class that all child nodes will be added to.
 # This is the root object for TipTap.
 module TipTap
-  class Document
-    include JsonRenderable
-    include HtmlRenderable
-    include PlainTextRenderable
-    include HasContent
-
+  class Document < Node
     self.type_name = "doc"
     self.html_tag = :div
     self.html_class_name = "tiptap-document"
