+++ lib/tip_tap/json_renderable.rb
@@ -4,25 +4,6 @@ require "tip_tap/registry"
 
 module TipTap
   module JsonRenderable
-    def self.included(base)
-      base.extend(ClassMethods)
-    end
-
-    module ClassMethods
-      def type_name=(type_name)
-        @type_name = type_name
-        Registry.register(type_name, self)
-      end
-
-      def type_name
-        @type_name
-      end
-    end
-
-    def type_name
-      self.class.type_name
-    end
-
     def include_empty_content_in_json?
       true
     end
